==> src/01-data-ops-preprocessing.ipynb <==
This notebook was used to perform pre-processing operations on the collected dataset from GitHub Google BigQuery Public Data. 

For details about source from where original data was collected, please refer to the `data/` directory at the root of this repository. The data used in this notebook was collected from [GitHub Google BigQuery Public Dataset](https://cloud.google.com/blog/topics/public-datasets/github-on-bigquery-analyze-all-the-open-source-code).

<hr>

At a high level, the data input from `data/*.csv` files have been saved into `dataset.json` file for later use in:
1. classification model
2. text retrieval

both of which form part of the project web UI.


import pandas as pd
import numpy as np
import random
import json
import regex as re

from IPython.display import display, Markdown

from sklearn.utils import shuffle
from dockerfile_parse import DockerfileParser

random.seed(10)
np.random.seed(10)

sh_df = pd.read_csv('../data/sample_contents_sh.csv')
bat_df = pd.read_csv('../data/sample_contents_bat.csv')
df_df = pd.read_csv('../data/sample_contents_dockerfile.csv')

sh_df.head()

bat_df.head()

df_df.head()

print(f'Dataset contains {len(sh_df)} shell scripts, {len(bat_df)} batch files and {len(df_df)} Dockerfiles.')
print(f'Thus, our corpus would be total of {len(sh_df) + len(bat_df) +  len(df_df)} documents.')

display(Markdown('## Let\'s look at a random .sh file:<br><br>\n\n```sh\n' + sh_df.loc[1020]['content'] + '\n```'))

display(Markdown('## Also, look at a random .bat file:<br><br>\n\n```sh\n' + bat_df.loc[1]['content'] + '\n```'))

display(Markdown('## And, how a typical Dockerfile would look like? Let\'s see: <br><br>\n\n```dockerfile\n' + df_df.loc[67]['content'] + '```'))

Parse the contents of a Dockerfile and produce sh-like contents.

def get_sh_content_from_dockerfile(content):
    dfp = DockerfileParser()
    dfp.content = content
    
    sh = ""
    for line in dfp.structure:
        if line['instruction'] == "FROM":
            sh += "export BASE_IMAGE=" + line["value"] + "\n"
        elif line['instruction'] == "RUN":
            sh += line["value"] + "\n"
        elif line['instruction'] == "ENV":
            sh += "export " + "=".join(line["value"].split(" ")) + "\n"
        elif line['instruction'] == "COPY":
            sh += "cp -r " + line["value"] + "\n"
        elif line['instruction'] == "CMD":
            try:
                sh += " ".join(json.loads(line["value"])) + "\n"
            except:
                sh += line["value"] + "\n"
        elif line['instruction'] == "ENTRYPOINT":
            try:
                sh += " ".join(json.loads(line["value"])) + "\n"
            except:
                sh += line["value"] + "\n"
        else:
            pass
    return sh

dfc = df_df.loc[67]['content']

parsh = get_sh_content_from_dockerfile(dfc)

display(Markdown(f"""
**Input**:

```dockerfile
{dfc}
```

<hr>

**Output**:
```sh
{parsh}
```

"""))

Creating extra column to store Dockerfile source and the parsed Dockerfile shell contents into the current content column.

df_df["original_content"] = df_df["content"]
df_df["content"] = df_df["original_content"].apply(get_sh_content_from_dockerfile)
df_df.head()

Combine and concat the contents of the 3 sources (contents from .sh, .bat, Dockerfile) into a single dataframe.

sh_df["original_content"] = ""
bat_df["original_content"] = ""

sh_df["source_category"] = '.sh'
df_df["source_category"] = 'Dockerfile'
bat_df["source_category"] = '.bat'

df = pd.concat([sh_df, df_df, bat_df])
df.reset_index(inplace=True)
df.drop(columns=['index'], inplace=True)

shuffle(df)
df.reset_index(inplace=True)
df.drop(columns=['index'], inplace=True)
df.head()

df.tail()

len(df)

Save the merged dataset to a CSV file so it can be re-used later.

df.to_json('./web/data/dataset.json')

Load the saved CSV dataset from file (just to verify contents).

df_again = pd.read_json('./web/data/dataset.json')
df_again.head()
