==> src/02-data-ops-classification-experiment.ipynb <==
**Alert: \[Experimental Approach\]** This approach was not used in the final project UI which was deployed. If you're looking for the final approach, consider seeing [other notebook: 03-data-ops-classification-final.ipynb](./03-data-ops-classification-final.ipynb) instead.

<hr>

This approach did not yield good results on the classification performance of the actual downstream task. It was identified that the root cause for such mis-classification was a result of manually created labels which were not robust enough to classify the data properly.

In this notebook, we use the pre-processed dataset prepared in the [previous notebook: 01-data-ops-preprocessing.ipynb](./01-data-ops-preprocessing.ipynb) and try to use it to do the following:

1. Manually label each shell script / batch file contents on the basis of gathered keywords
2. Score them using a Bag of Words based unigram token giving a reward for each keyword matched across corresponding labels
3. Aggregate the counts of the relevant keywords and select the label of the document based upon highest rewards for each category.
4. Split the dataset into 70:30 ratio for training and testing
5. Train the classifier using Bag of Words tokenization + TFIDF preprocessing in the pipeline followed by comparing results of SVM classifier and Naive Bayes classifier.

**Challenges (with RCA)**: As a result of the manually creation of labels, the classification data wasn't robust in terms of labelling causing classification failures for almost all queries. Also, the 'unknown' class was dominant in the classification problem and as a result of curse of imbalance faced by classifiers, it almost always will classify unknown which wasn't relevant.

import pandas as pd
from sklearn.utils import shuffle
import metapy
import numpy as np
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfTransformer
from sklearn.naive_bayes import MultinomialNB
from sklearn.svm import SVC
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report
from sklearn.pipeline import Pipeline

df = pd.read_json('./web/data/dataset.json')
df.head()

df_req = df.copy().drop(columns=['id', 'sample_repo_name', 'sample_path', 'original_content', 'size'])
df_req.head()

df_req = shuffle(df_req)
df_req.head()

keywords = {
    'rhel/fedora': [
        'systemctl',
        'useradd', 'passwd', 'userdel',
        'os', 'fedora', 'rhel', 'uname', 'cat',
        'podman', 'oc', 'yum', 'dnf',
        'microdnf', 'subscription', 'manager',
        'ubi', 'ip', 'journalctl', 'selinux'
    ],
    'ubuntu/debian': [
        'apt', 'get', 'dpkg',
        'debian', 'ubuntu', 'systemctl',
        'useradd', 'passwd', 'userdel',
        'lsb', 'uname', 'cat', 'lts', 'ifconfig',
        'ip'
    ],
    'windows': [
        'systeminfo', 'ver', 'choco',
        'sc', 'net', 'ipconfig', 'reg',
        'cmd', 'powershell', 'msiexec', 'chdir',
        'call', 'rmdir', 'move', 'cls', 'assoc', 'tasklist'
    ],
    'others/unknown': [
        'apk', 'pacman', 'alpine', 'arch',
    ]
}

token_set = set()

def preprocess(content):
    tok = metapy.analyzers.ICUTokenizer(suppress_tags=True)
    ngram = metapy.analyzers.NGramWordAnalyzer(1, tok)
    
    try:
        doc = metapy.index.Document()
        doc.content(content)
        unigrams = ngram.analyze(doc)
        
        tok.set_content(doc.content())
    except:
        return {}
    
    tokens_counts = {}
    for token, count in unigrams.items():
        tokens_counts[token] = count
        token_set.add(token)
    
    return tokens_counts

# manual labelling use gather keywords rewards
def label(contents, source_category):
    tc = preprocess(contents)
    ctr = {}
    for key in keywords:
        ctr[key] = 0
    
    for cat in keywords:
        for kw in keywords[cat]:
            if kw in tc:
                ctr[cat] += 1
    
    cat_names = np.array([key for key in ctr])
    cats = np.array([ctr[key] for key in cat_names])
    i0 = np.argmax(cats)
    
    cats1 = cats.copy()
    cats1[i0] = -1
    i1 = np.argmax(cats1)
    if ctr[cat_names[i0]] == ctr[cat_names[i1]]:
        if source_category == '.bat':
            return 'windows'
        else:
            return 'others/unknown'
    
    cat = cat_names[i0]
    return cat

labels = []
for _, row in df.iterrows():
    labels.append(label(row.content, row.source_category))

print('total tokens in corpus:', len(token_set))

ctr = {k: 0 for k in keywords}
for label in labels:
    ctr[label] += 1

print('dataset distribution:', ctr)

cats = {'rhel/fedora': 0, 'ubuntu/debian': 1, 'windows': 2, 'others/unknown': 3}
rev_cats = {0: 'rhel/fedora', 1: 'ubuntu/debian', 2: 'windows', 3: 'others/unknown'}
label_cats = [cats[label] for label in labels]

X = np.array(df_req.content)
y = np.array(label_cats)

# ignore empty contents
for i in range(len(X)):
    if X[i] is None:
        X[i] = ''

# train test split
X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.3, random_state=10)

SVM classifier

svm_clf = Pipeline([
    ('vect', CountVectorizer()),
    ('tfidf', TfidfTransformer()),
    ('clf', SVC()),
])

svm_clf.fit(X_train, y_train)

y_pred = svm_clf.predict(X_test)

print('SVM Classification Report:')
print(classification_report(y_test, y_pred))

# sample user queries

queries = [
    'sudo apt-get install git', # ground truth: rhel/fedora
    'cls rem rmdir', # windows
    'sudo dnf install -y code', # ubuntu/debian
    'apk add golang' # unknown/others
]

clfe = svm_clf.predict(np.array(queries))
print([{q: rev_cats[i]} for q, i in zip(queries, clfe)])

Naive Bayes classifier

nb_clf = Pipeline([
    ('vect', CountVectorizer()),
    ('tfidf', TfidfTransformer()),
    ('clf', MultinomialNB()),
])

nb_clf.fit(X_train, y_train)

y_pred = nb_clf.predict(X_test)

print('Naive Bayes Classification Report:')
print(classification_report(y_test, y_pred))

# sample user query
clfe = nb_clf.predict(np.array(queries))
print([{q: rev_cats[i]} for q, i in zip(queries, clfe)])
